Move EventSource from GlobalStore to Server
The Server can generate a multitude of events, not limited to just
global change events. It doesn't make sense to store event slots and
listeners in the GlobalStore.

The initial globals list event is used to be sent by
GlobalStore::add_listener, it's not sent by the post-insert operation of
globals::Registry::bind