apollo: xdg_shell: send configure events
The clients are required by the protocol to wait for a
xdg_surface.configure event before attaching a buffer to a surface. So
we have a send this event to unblock clients like eglgears. (Some
non-conforming clients don't wait.)

Add new API to allow a Shell impl to report the layout it has for a
window, so we can send it in the configure event sequence.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>