Replace trait objects with enum
The current limitation of trait objects in Rust is quite annoying. And
we generates a big match with all the object types anyway, might as well
go all in.

Also move DefaultShell into crescent. The Shell are supposed to be
implemented by the compositor author, so make sure that's actually
possible.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>