wl_server: require user to explicitly list object types
Instead of indirectly by generating a enum from the union of
`<Global as Bind>::Objects`, in macros. This make the type simpler, and
we will be able to request `Object: From<TheObjectWeNeed>`, instead of
`Object: From<ACollectionOfObjects>`.

Also add a `dispatch` function to ClientContext, and move the macro for
implementing that out of `globals!`.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>