wl_server: make the object store shareable
Now the object store can be operated with a shared reference. This is a
preparation for running multiple async tasks with a client context.

And object store no longer store Rc<Object>, instead we make object
methods not take `&self`. Reminder, this and the previous Rc<Object>
trick is used so we don't borrow the object store when calling the
object methods, so the object methods can modify the object store. I
think this is a better solution than using Rc<Object>, one advantage is
this way we aren't forcing an `Rc` on the user. We will see if this
actually make a difference.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>