apollo: don't store frame callbacks per surface state
Realize each state's frame callback list is a prefix of all frame
callbacks ever added to any states of the corresponding surface. So we
can store the list of callbacks in the Surface, instead of the
SurfaceState, and only keep an index in SurfaceState.

This way we don't need to track all states of a surface for frame
callback deletion - we just shift the list in Surface. We also don't
need to clone the callback list when cloneing SurfaceState.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>