New event handling mechanism
This is maybe take 3 or take 4 of how event handling works. This maybe
premature celebration, but I have the feeling this time it is pretty
close to its final form.

This is based on the EventSource insight we had during the last event
handling refactor. With it we don't need to generate a event handle from
the client context, which required uses of macro and trait magic to
allocate a unique id at compile time for each type that needs event
delievery. That was hard to explain and inflexible. With EventSource,
you can get and a stream and register it with the client context with a
event handler. This has the best of both worlds: it's easier to
understand for the user; and the client tasks don't have to share their
client contexts with other tasks. The cost is some runtime overhead
that should be negligible, and slightly more difficulty in writing the
event handlers, because you have to write them in a poll style, without
the use of async functions.

Next we can probably change accesses to object store to use a unique
reference, and remove all those clones.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>