wl_server: allow use of async {} for EventHandler
Before EventHandlers are implemented with poll style functions, this is
because we need to type erase these EventHandlers, if they returns a
future, those future would have to be boxed, and we don't want these
allocations.

In this commit, we use a convoluted (from our perspective, this should
be as simple for the user) trick to solve this problem. For each
EventHandler, we have a preallocated space to store its future, packed
into the same type erased Box as the EventHandler. Only problem is when
these box don't contain a future, we need them to be `'static`, so we
can easily store them, and when they contain a future, they would have
to capture the future's lifetime. We use a small `unsafe {}` block to
change the lifetime of the box back and forth. And now we can do all the
nice things we can do in a `async fn`!

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>