wl_server: add singleton states associated to object types
A common pattern that emerged is having a state shared between objects
of the same type. So add the machinary to make this easier. This should
make the use of Rc<RefCell<>> unnecessary.

And Store::get now also does `.cast` for you. Merge these 2 commonly
done together operations together, and improve error handling a little
bit.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>