wl_server: divorce AnyObject and MonoObject traits from Object
An Object can be _either_ an AnyObject or a MonoObject, so neither of
them can be the supertrait. We now explicitly differentiate AnyObject
from MonoObject to avoid possible confusion. Previously a MonoObject
is also an AnyObject, which is technically correct but made things
harder to explain.

Also added chunks of doc comments.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>