apollo: refactor wl_buffer
Restructure the types and generics so code reuse is a bit easier.

Fix the buffer release logic. Previously the release event is sent when
a surface commit changed the attached buffer. Which is totally wrong.
For one, the buffer might still be used in an older surface state that
is currently visible because of synchronized state update; and the
compositor might want to release a buffer earlier, when it's still
attached to a surface.

We made buffers event sources, so the compositor could trigger a release
event at any point. We also keep track of how many surface state is
using a buffer, so we can automatically release when that number drops
to 0.

Know bug:

If the client re-attach the same buffer that was attached to the
surface, after we have sent a release event for that buffer, we are
again allowed to read that buffer's content. So we should clear the
released flag on the buffer. But we don't because we only clear the flag
when the attached buffer changes.

We need to fix our surface pending state and commit logic.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>