wl_server: don't wrap singleton states in Option
Use unit instead for types that don't need it.

This was done to avoid storing anything for types that don't have a
singleton state. But saving a couple dozen bytes per object type isn't
really worth the trouble. Besides, now we can also quickly check if
objects of a certain type exist in the store by trying to get its
singleton state.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>