runa-orbiter/shell: fix handling of surface updates
A surface's pending updates should not be represented as a "shadow
state". i.e. it's not a full copy of the surface state, and changes
apply to it like they would to a normal surface state. Instead it stores
what the new states _should be_. And during commit, these new states are
applied in a predefined order. All these except for the surface stack,
which _does_ actually have a shadow stack stored in pending updates.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>