runa-orbiter/shell: be less clever about frame callbacks management
We used to store frame callbacks in Surface, which means all
SurfaceStates can share storage for frame callbacks. But this makes
management of Surface lifetime annoying, and has some index overflow
problems.

This commit makes storage of frame callbacks straightforward. This takes
a bit more memory but should be simpler.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>